% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINLAM.R
\name{dq_Z}
\alias{dq_Z}
\title{Compute density transition probability 
Created by Virgilio Gómez-Rubio, https://becarioprecario.bitbucket.io/inla-gitbook/ch-mixture.html}
\usage{
dq_Z(z.old, z.new, log = TRUE)
}
\arguments{
\item{z.old}{initial vector containing subpopulation/subgroup membership (z)}

\item{z.new}{transitioned vector of subpopulation/subgroup membership (by inla)}
}
\value{
vector
}
\description{
Compute density transition probability 
Created by Virgilio Gómez-Rubio, https://becarioprecario.bitbucket.io/inla-gitbook/ch-mixture.html
}
