% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINLAM.R
\name{fit_inla}
\alias{fit_inla}
\title{Wrapper for INLA model fitting
Created by Virgilio Gómez-Rubio, https://becarioprecario.bitbucket.io/inla-gitbook/ch-mixture.html}
\usage{
fit_inla(y, grp)
}
\arguments{
\item{y}{numeric, dependent variable to predict}

\item{grp}{numeric, number of subpopulations/subgroups}
}
\value{
vector
}
\description{
Wrapper for INLA model fitting
Created by Virgilio Gómez-Rubio, https://becarioprecario.bitbucket.io/inla-gitbook/ch-mixture.html
}
