% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINLAM.R
\name{fit_inla_internal}
\alias{fit_inla_internal}
\title{Created by Virgilio Gómez-Rubio, https://becarioprecario.bitbucket.io/inla-gitbook/ch-mixture.html}
\usage{
fit_inla_internal(y, grp, prior_means)
}
\arguments{
\item{y}{numeric, dependent variable to predict}

\item{grp}{numeric, estimated subpopulation/subgroup number}

\item{prior_means}{numeric, internally defined prior means of each subgroup}
}
\value{
data frame
}
\description{
Created by Virgilio Gómez-Rubio, https://becarioprecario.bitbucket.io/inla-gitbook/ch-mixture.html
}
