% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINLAM.R
\name{fuss_PARALLEL}
\alias{fuss_PARALLEL}
\title{Parallelizaton FRAME for the MAIN function}
\usage{
fuss_PARALLEL(
  data,
  varCLASS,
  varY,
  method = "dpi",
  within = 0.03,
  maxNGROUP = 5,
  df_prob = FALSE,
  out_dir = NULL,
  n_workers = 4
)
}
\arguments{
\item{data}{input data}

\item{varCLASS}{character, variable/column (categorical, character or numeric) name containing the subpopulations/subgroups}

\item{varY}{character, variable/column (numeric) containing the values assigned to the population/subpopulations}

\item{method}{character, which density estimator to apply: "nrd", "bcv", "dpi", by default "dpi", see above}

\item{within}{numeric, range as in group_MODES()}

\item{maxNGROUP}{numeric, after data inspection -> assumed maximum number of subpopulations/subgroups}

\item{df_prob}{logical, as in get_weighted_probs()}

\item{out_dir}{by default NULL, if character, path of the output directory to write csv tables}

\item{n_workers}{numeric, number of cores to run simulations on, by default 4}
}
\value{
data frame or list of data frames (stored as an R object or written as csvs)
}
\description{
Parallelizaton FRAME for the MAIN function
}
