% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINLAM.R
\name{get_mode}
\alias{get_mode}
\title{Compute mode based on density estimation}
\usage{
get_mode(x)
}
\arguments{
\item{x}{vector, samples of a distribution}
}
\value{
numeric
}
\description{
Compute mode based on density estimation
}
