% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINLAM.R
\name{get_MODES}
\alias{get_MODES}
\title{Extract mode statistics from a mode forest}
\usage{
get_MODES(res_modeforest, nmod)
}
\arguments{
\item{res_modeforest}{multimode::modeforest() result}

\item{nmod}{numeric, number of subpopulations/subgroups}
}
\value{
data frame
}
\description{
Extract mode statistics from a mode forest
}
