% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINLAM.R
\name{get_PROBCLASS_MH}
\alias{get_PROBCLASS_MH}
\title{MAIN function}
\usage{
get_PROBCLASS_MH(
  data,
  varCLASS,
  varY,
  method = "dpi",
  within = 0.03,
  maxNGROUP = 5,
  df_prob = FALSE,
  out_dir = NULL
)
}
\arguments{
\item{data}{input data}

\item{varCLASS}{character, variable/column (categorical, character or numeric) name containing the subpopulations/subgroups}

\item{varY}{character, variable/column (numeric) containing the values assigned to the population/subpopulations}

\item{method}{character, which density estimator to apply: "nrd", "bcv", "dpi", by default "dpi", see above}

\item{within}{numeric, range as in get_MODES()}

\item{maxNGROUP}{numeric, after data inspection -> assumed maximum number of subpopulations/subgroups}

\item{df_prob}{logical, as in get_weighted_probs()}

\item{out_dir}{by default NULL, if character, path of the output directory to write csv tables}
}
\value{
data frame or list of data frames (stored as an R object or written as csvs)
}
\description{
"dpi" ~ Direct plug-in by Sheather & Jones, 
"bcv" ~ Biased cross validation by Jonas and Kappenman (1991)
"nrd" ~ Scott's 1.06 (modified Silvermann's rule of thumb 1.34)
}
