% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINLAM.R
\name{get_probs}
\alias{get_probs}
\title{Compute probabilities for group membership
Created by Virgilio Gómez-Rubio, https://becarioprecario.bitbucket.io/inla-gitbook/ch-mixture.html}
\usage{
get_probs(z)
}
\arguments{
\item{z}{vector of subpopulation/subgroup membership}
}
\value{
numeric vector
}
\description{
Compute probabilities for group membership
Created by Virgilio Gómez-Rubio, https://becarioprecario.bitbucket.io/inla-gitbook/ch-mixture.html
}
