% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINLAM.R
\name{get_weighted_probs}
\alias{get_weighted_probs}
\title{Get the weights based on probabilities of class-belonging}
\usage{
get_weighted_probs(get_z_object, y, grp, main_class, df_prob = FALSE)
}
\arguments{
\item{get_z_object}{get_Z() result}

\item{y}{numeric, dependent variable to predict}

\item{grp}{numeric, number of subpopulations/subgroups}

\item{main_class}{numeric, subpopulation/subgroup id}

\item{df_prob}{by default FALSE (only summary), whether the estimated samples' probabilities to output}
}
\value{
data frame or list of data farmes
}
\description{
Get the weights based on probabilities of class-belonging
}
