% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINLAM.R
\name{inla_MH}
\alias{inla_MH}
\title{INLABMA::INLAMH (Metropolitan-Hastings sampling algorithm)
Created by Virgilio Gómez-Rubio, https://becarioprecario.bitbucket.io/inla-gitbook/ch-mixture.html
Source: https://rdrr.io/rforge/INLABMA/src/R/INLAMH.R}
\usage{
inla_MH(
  d,
  fit.inla,
  b.init,
  rq,
  dq,
  prior,
  n.sim = 200,
  n.burnin = 100,
  n.thin = 1,
  n.errors = 20,
  verbose = FALSE
)
}
\arguments{
\item{d}{data to be passed to inla() in a suitable format}

\item{fit.inla}{see above}

\item{b.init}{see above, nitial values of the model parameters}

\item{rq}{output of rq_Z(), sample from the proposal distribution}

\item{dq}{output of dq_Z(), density of the proposal distribution}

\item{prior}{prior on the model parameters}

\item{n.sim}{number of simulations}

\item{n.thin}{thinning after burn in}

\item{n.errors}{total number of allowed INLA errors}

\item{verbose}{by default FALSE, if TRUE information messages are printed}

\item{n.burin}{number simulations for burn in}
}
\value{
list of simulations
}
\description{
INLABMA::INLAMH (Metropolitan-Hastings sampling algorithm)
Created by Virgilio Gómez-Rubio, https://becarioprecario.bitbucket.io/inla-gitbook/ch-mixture.html
Source: https://rdrr.io/rforge/INLABMA/src/R/INLAMH.R
}
