% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINLAM.R
\name{prior_Z}
\alias{prior_Z}
\title{Prior probability of z
Created by Virgilio Gómez-Rubio, https://becarioprecario.bitbucket.io/inla-gitbook/ch-mixture.html}
\usage{
prior_Z(z, log = TRUE)
}
\arguments{
\item{z}{initial vector containing subpopulation/subgroup membership}

\item{log}{logical, by default TRUE}
}
\value{
vector
}
\description{
Prior probability of z
Created by Virgilio Gómez-Rubio, https://becarioprecario.bitbucket.io/inla-gitbook/ch-mixture.html
}
