% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINLAM.R
\name{rq_Z}
\alias{rq_Z}
\title{Random group assignment generator
Created by Virgilio Gómez-Rubio, https://becarioprecario.bitbucket.io/inla-gitbook/ch-mixture.html}
\usage{
rq_Z(z)
}
\arguments{
\item{z}{initial vector containing subpopulation/subgroup membership (proposed distribution)}
}
\value{
list
}
\description{
Random group assignment generator
Created by Virgilio Gómez-Rubio, https://becarioprecario.bitbucket.io/inla-gitbook/ch-mixture.html
}
